requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
djangorestframework-simplejwt==5.3.0
django-rest-swagger==2.2.0
web3==6.11.1
pytest==7.4.3
pytest-django==4.5.2